refactor: remove unused code, organize imports
