+++ .github/workflows/monthly_roadmap_reminder.yml
@@ -34,5 +34,5 @@ jobs:
         with:
           github-token: ${{ secrets.token }}
           script: |
-            const { default: main } = await import('${{ github.workspace }}/packages/monthly_roadmap/src/index.mjs')
+            const { default: main } = await import('${{ github.workspace }}/packages/reporting/src/index.mjs')
             await main({github, context, core})
