+++ packages/github/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "github",
-  "version": "1.0.0",
-  "description": "Powertools GitHub client",
-  "devDependencies": {}
+	"name": "github",
+	"version": "1.0.0",
+	"description": "Powertools GitHub client",
+	"devDependencies": {}
 }
