+++ packages/github/src/client/index.mjs
@@ -1,0 +1,1 @@
+export { GitHub } from "./GitHub.mjs";
