+++ packages/github/tests/contract/update_issue.test.mjs
@@ -1,8 +1,8 @@
-import {setupServer} from "msw/node";
-import {buildIssues} from "testing/src/builders/issues.mjs";
-import {updateIssueFailureHandler, updateIssueHandler} from "testing/src/interceptors/issues_handler.mjs";
-import {afterAll, afterEach, beforeAll, describe, expect, it} from "vitest";
-import {GitHub} from "../../src/client/GitHub.mjs";
+import { GitHub } from "github/src/client";
+import { setupServer } from "msw/node";
+import { buildIssues } from "testing/src/builders";
+import { updateIssueFailureHandler, updateIssueHandler } from "testing/src/interceptors/issues_handler.mjs";
+import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest";
 
 describe("update issues contract", () => {
 	process.env.GITHUB_REPOSITORY = "test-org/test-repo";
