+++ packages/github/tests/unit/create_or_update_issue.test.mjs
@@ -1,32 +1,30 @@
 import { faker } from "@faker-js/faker";
-import { buildIssues, buildSearchIssues } from "testing/src/builders/issues.mjs";
+import { GitHub } from "github/src/client";
+import { buildIssues } from "testing/src/builders";
 import { describe, expect, it, vi } from "vitest";
-import { GitHub } from "../../src/client/GitHub.mjs";
 
 describe("create or update issue", () => {
+	it("find issue using search", async () => {
+		// GIVEN
+		const existingIssue = buildIssues({ max: 1 });
+		const searchQuery = faker.lorem.sentence();
 
-    it("find issue using search", async () => {
-        // GIVEN
-        const existingIssue = buildIssues({ max: 1 });
-        const searchQuery = faker.lorem.sentence();
+		const github = new GitHub();
+		const findIssueSpy = vi.spyOn(github, "findIssue");
+		findIssueSpy.mockImplementation(() => existingIssue);
 
-        const github = new GitHub();
-        const findIssueSpy = vi.spyOn(github, "findIssue");
-        findIssueSpy.mockImplementation(() => existingIssue);
+		// WHEN
+		const issueFound = await github.findIssue({ searchQuery });
 
-        // WHEN
-        const issueFound = await github.findIssue({searchQuery})
-
-        // THEN
-        expect(issueFound).toBe(existingIssue);
-        expect(findIssueSpy).toHaveBeenCalledWith({
-            searchQuery,
-        });
-
-    })
+		// THEN
+		expect(issueFound).toBe(existingIssue);
+		expect(findIssueSpy).toHaveBeenCalledWith({
+			searchQuery,
+		});
+	});
 	it("create issue if one is not found", async () => {
 		// GIVEN
-		const searchResult = []
+		const searchResult = [];
 		const createdIssue = buildIssues({ max: 1 });
 		const searchQuery = faker.lorem.sentence();
 		const options = {
