+++ packages/reporting/src/issues/index.mjs
@@ -1,0 +1,5 @@
+export {
+	getTopOldestIssues,
+	getTopMostCommented,
+	getTopFeatureRequests,
+} from "./functions.mjs";
