+++ packages/reporting/src/models/index.mjs
@@ -1,0 +1,7 @@
+export { HighlyCommentedIssue } from "./HighlyCommentedIssue.mjs";
+
+export { LongRunningPullRequest } from "./LongRunningPullRequest.mjs";
+
+export { PopularFeatureRequest } from "./PopularFeatureRequest.mjs";
+
+export { OldestIssue } from "./OldestIssue.mjs";
