+++ packages/reporting/src/pull_requests/index.mjs
@@ -1,0 +1,1 @@
+export { getLongRunningPRs } from "./functions.mjs";
