+++ packages/testing/src/builders/reporting.mjs
@@ -1,8 +1,9 @@
 import { issueSchema } from "github/src/schemas/issues.mjs";
-import { HighlyCommentedIssue } from "monthly_roadmap/src/models/HighlyCommentedIssue.mjs";
-import { LongRunningPullRequest } from "monthly_roadmap/src/models/LongRunningPullRequest.mjs";
-import { OldestIssue } from "monthly_roadmap/src/models/OldestIssue.mjs";
-import { PopularFeatureRequest } from "monthly_roadmap/src/models/PopularFeatureRequest.mjs";
+import { HighlyCommentedIssue } from "reporting/src/models/HighlyCommentedIssue.mjs";
+import { LongRunningPullRequest } from "reporting/src/models/LongRunningPullRequest.mjs";
+import { OldestIssue } from "reporting/src/models/OldestIssue.mjs";
+import { PopularFeatureRequest } from "reporting/src/models/PopularFeatureRequest.mjs";
+import { z } from "zod"
 
 /**
  * Builds an array of top feature requests from GitHub issues
